requirements.txt
blinker==1.8.2
bokeh==3.1.1
boto3==1.34.100
botocore==1.34.100
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
et-xmlfile==1.1.0
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
tornado==6.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
werkzeug==3.0.3
XlsxWriter==3.2.0
xyzservices==2024.4.0
zipp==3.18.1
